Move runtime config to its own module
